+++ FEXCore/include/FEXCore/Utils/StringUtils.h
